+++ LearnComputer/LearnComputer/CircuitInfrustructure/Basic/IOutputEndpoint.cs
@@ -4,7 +4,7 @@ namespace LearnComputer.CircuitInfrustructure
 {
     public interface IOutputEndpoint
     {
-        Byte LastSentSignal { get; }
-        void Produce(Byte signal);
+        Int32 LastSentSignal { get; }
+        void Produce(Int32 signal);
     }
 }
